safari-534.53-branch/LayoutTests/platform/mac/fast/loader/file-url-mimetypes-3-expected.txt

example-s.s has MIME type text/plain
example-scpt.scpt has MIME type application/octet-stream
example-sd2.sd2 has MIME type application/octet-stream
example-sf2.sf2 has MIME type application/octet-stream
example-sgi.sgi has MIME type image/sgi
example-sgm.sgm has MIME type text/sgml
example-sgml.sgml has MIME type text/sgml
example-sh.sh has MIME type application/x-sh
example-shar.shar has MIME type application/x-shar
example-shtm.shtm has MIME type text/html
example-shtml.shtml has MIME type text/html
example-silo.silo has MIME type model/mesh
example-sit.sit has MIME type application/x-stuffit
example-sitx.sitx has MIME type application/x-stuffitx
example-skd.skd has MIME type application/x-koan
example-skm.skm has MIME type application/x-koan
example-skp.skp has MIME type application/x-koan
example-skt.skt has MIME type application/x-koan
example-smf.smf has MIME type audio/midi
example-smi.smi has MIME type application/smil
example-smil.smil has MIME type application/smil
example-snd.snd has MIME type audio/basic
example-so.so has MIME type application/octet-stream
example-spl.spl has MIME type application/x-futuresplash
example-src.src has MIME type application/x-wais-source
example-srf.srf has MIME type application/octet-stream
example-suit.suit has MIME type application/octet-stream
example-sv4cpio.sv4cpio has MIME type application/x-sv4cpio
example-sv4crc.sv4crc has MIME type application/x-sv4crc
example-svg.svg has MIME type image/svg+xml
example-svgz.svgz has MIME type image/svg+xml
example-swf.swf has MIME type application/x-shockwave-flash
example-t.t has MIME type application/x-troff
example-tar.tar has MIME type application/x-tar
example-targa.targa has MIME type image/x-targa
example-tbz.tbz has MIME type application/x-bzip2
example-tbz2.tbz2 has MIME type application/x-bzip2
example-tcl.tcl has MIME type application/x-tcl
example-tex.tex has MIME type application/x-tex
example-texi.texi has MIME type application/x-texinfo
example-texinfo.texinfo has MIME type application/x-texinfo
example-text.text has MIME type text/plain
example-tga.tga has MIME type image/targa
example-tgz.tgz has MIME type application/x-gzip
example-tif.tif has MIME type image/tiff
example-tiff.tiff has MIME type image/tiff
example-torrent.torrent has MIME type application/x-bittorrent
example-tr.tr has MIME type application/x-troff
example-tsv.tsv has MIME type text/tab-separated-values
example-ttc.ttc has MIME type application/octet-stream
example-ttf.ttf has MIME type application/octet-stream
example-txt.txt has MIME type text/plain
example-ul.ul has MIME type application/octet-stream
example-ulaw.ulaw has MIME type application/octet-stream
example-ulw.ulw has MIME type application/octet-stream
example-ustar.ustar has MIME type application/x-ustar
example-vcard.vcard has MIME type text/directory
example-vcd.vcd has MIME type application/x-cdlink
example-vcf.vcf has MIME type text/directory
example-vfw.vfw has MIME type video/avi
example-vrml.vrml has MIME type model/vrml
example-war.war has MIME type application/java-archive
example-wav.wav has MIME type audio/wav
example-wave.wave has MIME type audio/wav
example-wax.wax has MIME type video/x-ms-wax
example-wbmp.wbmp has MIME type image/vnd.wap.wbmp
example-wbxml.wbxml has MIME type application/vnd.wap.wbxml
example-webarchive.webarchive has MIME type application/x-webarchive
example-wm.wm has MIME type video/x-ms-wm
example-wma.wma has MIME type video/x-ms-wma
example-wmd.wmd has MIME type application/x-ms-wmd
example-wml.wml has MIME type text/vnd.wap.wml
example-wmlc.wmlc has MIME type application/vnd.wap.wmlc
example-wmls.wmls has MIME type text/vnd.wap.wmlscript
example-wmlsc.wmlsc has MIME type application/vnd.wap.wmlscriptc
example-wmp.wmp has MIME type video/x-ms-wmp
example-wmv.wmv has MIME type video/x-ms-wmv
example-wmx.wmx has MIME type video/x-ms-wmx
example-wrl.wrl has MIME type model/vrml
example-wvx.wvx has MIME type video/x-ms-wvx
example-xbm.xbm has MIME type image/x-xbitmap
example-xdp.xdp has MIME type application/vnd.adobe.xdp+xml
example-xfd.xfd has MIME type application/vnd.adobe.xfd+xml
example-xfdf.xfdf has MIME type application/vnd.adobe.xfdf
example-xht.xht has MIME type application/xhtml+xml
example-xhtm.xhtm has MIME type application/xhtml+xml
example-xhtml.xhtml has MIME type application/xhtml+xml
example-xls.xls has MIME type application/vnd.ms-excel
example-xlsb.xlsb has MIME type application/vnd.ms-excel.sheet.binary.macroenabled.12
example-xlsm.xlsm has MIME type application/vnd.ms-excel.sheet.macroenabled.12
example-xlsx.xlsx has MIME type application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
example-xltm.xltm has MIME type application/vnd.ms-excel.template.macroenabled.12
example-xltx.xltx has MIME type application/vnd.openxmlformats-officedocument.spreadsheetml.template
example-xml.xml has MIME type application/xml
example-xpm.xpm has MIME type image/x-xpixmap
example-xsl.xsl has MIME type text/xml
example-xwd.xwd has MIME type image/x-xwindowdump
example-xyz.xyz has MIME type chemical/x-xyz
example-y.y has MIME type text/plain
example-ym.ym has MIME type text/plain
example-ymm.ymm has MIME type text/plain
example-ypp.ypp has MIME type text/plain
example-yxx.yxx has MIME type text/plain
example-zip.zip has MIME type application/zip
Using an image element to load resources, this test makes sure that every MIME type the loader should know about based solely on file extension in file:// urls is correct.
